Move UnassignedVariableUsageInspectionTests for QuickFixes into relevant files under RD.QuickFixes
