+++ RubberduckTests/Inspections/UnassignedVariableUsageInspectionTests.cs
@@ -128,65 +128,6 @@ End Sub";
 
             Assert.IsFalse(inspectionResults.Any());
         }
-        
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void UnassignedVariableUsage_QuickFixWorks()
-        {
-            const string inputCode =
-@"Sub Foo()
-    Dim b As Boolean
-    Dim bb As Boolean
-    bb = b
-End Sub";
-
-            const string expectedCode =
-@"Sub Foo()
-    Dim b As Boolean
-    Dim bb As Boolean
-    
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new UnassignedVariableUsageInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-
-            new RemoveUnassignedVariableUsageQuickFix(state).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
-
-        [TestMethod]
-        [TestCategory("Inspections")]
-        public void UnassignedVariableUsage_IgnoreQuickFixWorks()
-        {
-            const string inputCode =
-@"Sub Foo()
-    Dim b As Boolean
-    Dim bb As Boolean
-    bb = b
-End Sub";
-
-            const string expectedCode =
-@"Sub Foo()
-    Dim b As Boolean
-    Dim bb As Boolean
-'@Ignore UnassignedVariableUsage
-    bb = b
-End Sub";
-
-            IVBComponent component;
-            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
-            var state = MockParser.CreateAndParse(vbe.Object);
-
-            var inspection = new UnassignedVariableUsageInspection(state);
-            var inspectionResults = inspection.GetInspectionResults();
-            
-            new IgnoreOnceQuickFix(state, new[] {inspection}).Fix(inspectionResults.First());
-            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
-        }
 
         [TestMethod]
         [TestCategory("Inspections")]
