+++ RubberduckTests/QuickFixes/RemoveUnassignedVariableUsageQuickFix.cs
@@ -1,0 +1,43 @@
+﻿using System.Linq;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Rubberduck.Inspections.Concrete;
+using Rubberduck.Inspections.QuickFixes;
+using Rubberduck.VBEditor.SafeComWrappers.Abstract;
+using RubberduckTests.Mocks;
+
+namespace RubberduckTests.QuickFixes
+{
+    [TestClass]
+    public class RemoveUnassignedVariableUsageQuickFixTests
+    {
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void UnassignedVariableUsage_QuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Foo()
+    Dim b As Boolean
+    Dim bb As Boolean
+    bb = b
+End Sub";
+
+            const string expectedCode =
+@"Sub Foo()
+    Dim b As Boolean
+    Dim bb As Boolean
+    
+End Sub";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new UnassignedVariableUsageInspection(state);
+            var inspectionResults = inspection.GetInspectionResults();
+
+            new RemoveUnassignedVariableUsageQuickFix(state).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+    }
+}
